Buttons!
Reworking the way that buttons are used on the site. Making better use of a reusable component, allowing for better scalability and easier implementation. 

Created a hashlink component to allow for desired UX effect when clicking on navigation buttons that will auto scroll to the correct element in the page.