App Loading State
Created app state to hold the value of the app being ready (based upon the large images being fully loaded on the page) and the animation for the fade in/fade out being complete. 

Also created an array that contains all the images that require loading before displaying the app. This will allow us to be able to dynamically change the amount of images/assets we are requiring to load.

In total the application will take at minimum 1.7 seconds to display the main content. The main adjustment for time should be the first timeout set at 1000.